rename drizzle-db to db
