+++ apps/server/src/functions/cron/scrapeEventsHandler.ts
@@ -1,4 +1,3 @@
-import { createEvents, scrapeEvents } from "@zotmeal/api";
-import { PrismaClient } from "@zotmeal/db";
+import { upsertEvents, scrapeEvents } from "@zotmeal/api";
 
-export const main = async (event, context) => {};
+export const main = async (event, context) => { };
