+++ package.json
@@ -11,7 +11,7 @@
     "clean": "git clean -xdf node_modules",
     "clean:workspaces": "turbo clean",
     "db:push": "pnpm -F db push",
-    "db:studio": "pnpm --filter drizzle-db run db:studio",
+    "db:studio": "pnpm --filter db run db:studio",
     "dev": "turbo dev --parallel --concurrency=13",
     "format": "turbo format --continue -- --cache --cache-location node_modules/.cache/.prettiercache",
     "format:fix": "turbo format --continue -- --write --cache --cache-location node_modules/.cache/.prettiercache",
