+++ packages/api/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@trpc/server": "11.0.0-next-beta.308",
-    "@zotmeal/drizzle-db": "workspace:^",
+    "@zotmeal/db": "workspace:^",
     "@zotmeal/eslint-config": "workspace:^",
     "@zotmeal/prettier-config": "workspace:^",
     "@zotmeal/tsconfig": "workspace:^",
