--- packages/db/index.ts
@@ -1,1 +1,0 @@
-export * from "@prisma/client";
