+++ packages/db/package.json
@@ -1,30 +1,34 @@
 {
   "name": "@zotmeal/db",
-  "version": "0.1.0",
+  "version": "0.0.0",
   "private": true,
   "type": "module",
   "license": "MIT",
-  "main": "./index.ts",
-  "types": "./index.ts",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
   "scripts": {
-    "clean": "rm -rf .turbo node_modules",
-    "format": "prettier --check . --ignore-path ../../.gitignore",
-    "lint": "eslint .",
-    "db:generate": "pnpm with-env prisma generate",
-    "db:push": "pnpm with-env prisma db push --skip-generate",
-    "db:studio": "pnpm with-env prisma studio",
-    "typecheck": "tsc --noEmit",
+    "dev": "echo 'Add dev script here'",
+    "build": "echo 'Add build script here'",
+    "test": "echo 'Add test script here'",
+    "lint": "echo 'Add lint script here'",
+    "db:generate": "drizzle-kit generate:pg --schema=./src/schema --out=./migrations",
+    "db:push": "pnpm with-env drizzle-kit push:pg",
+    "db:migrate": "pnpm with-env tsx ./src/migrate.ts",
+    "db:studio": "pnpm with-env drizzle-kit studio --port 3000 --verbose",
+    "db:drop": "pnpm with-env drizzle-kit drop",
+    "db:check": "pnpm with-env drizzle-kit check:pg",
     "with-env": "dotenv -e ../../.env --"
   },
   "devDependencies": {
+    "@types/pg": "^8.11.2",
     "@zotmeal/eslint-config": "workspace:^0.2.0",
     "@zotmeal/prettier-config": "workspace:^0.1.0",
     "@zotmeal/tsconfig": "workspace:^0.1.0",
     "dotenv-cli": "^7.3.0",
+    "drizzle-kit": "^0.20.14",
     "eslint": "^8.57.0",
     "prettier": "^3.2.5",
-    "typescript": "^5.3.3",
-    "prisma": "^5.10.2"
+    "typescript": "^5.3.3"
   },
   "eslintConfig": {
     "root": true,
@@ -34,7 +38,11 @@
   },
   "prettier": "@zotmeal/prettier-config",
   "dependencies": {
-    "@prisma/client": "^5.10.2",
-    "ts-node": "^10.9.2"
+    "@neondatabase/serverless": "^0.9.0",
+    "drizzle-orm": "^0.29.4",
+    "drizzle-zod": "^0.5.1",
+    "pg": "^8.11.3",
+    "tsx": "^4.7.1",
+    "zod": "^3.22.4"
   }
 }
