+++ packages/db/src/schema/station.ts
