--- packages/drizzle-db/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@zotmeal/drizzle-db",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "license": "MIT",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "scripts": {
-    "dev": "echo 'Add dev script here'",
-    "build": "echo 'Add build script here'",
-    "test": "echo 'Add test script here'",
-    "lint": "echo 'Add lint script here'",
-    "db:generate": "drizzle-kit generate:pg --schema=./src/schema --out=./migrations",
-    "db:push": "pnpm with-env drizzle-kit push:pg",
-    "db:migrate": "pnpm with-env tsx ./src/migrate.ts",
-    "db:studio": "pnpm with-env drizzle-kit studio --port 3000 --verbose",
-    "db:drop": "pnpm with-env drizzle-kit drop",
-    "db:check": "pnpm with-env drizzle-kit check:pg",
-    "with-env": "dotenv -e ../../.env --"
-  },
-  "devDependencies": {
-    "@types/pg": "^8.11.2",
-    "@zotmeal/eslint-config": "workspace:^0.2.0",
-    "@zotmeal/prettier-config": "workspace:^0.1.0",
-    "@zotmeal/tsconfig": "workspace:^0.1.0",
-    "dotenv-cli": "^7.3.0",
-    "drizzle-kit": "^0.20.14",
-    "eslint": "^8.57.0",
-    "prettier": "^3.2.5",
-    "typescript": "^5.3.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "@zotmeal/eslint-config/base"
-    ]
-  },
-  "prettier": "@zotmeal/prettier-config",
-  "dependencies": {
-    "@neondatabase/serverless": "^0.9.0",
-    "drizzle-orm": "^0.29.4",
-    "drizzle-zod": "^0.5.1",
-    "pg": "^8.11.3",
-    "tsx": "^4.7.1",
-    "zod": "^3.22.4"
-  }
-}
