--- packages/drizzle-db/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@zotmeal/tsconfig/base.json",
-  "compilerOptions": {
-    "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
-  },
-  "exclude": ["node_modules"]
-}
